+++ HISTORY.md
@@ -1,2 +1,7 @@
-- 0.0.1
+- 0.0.2 2013-09-06
+  - Working get, set, del.
+  - Using couchnode master branch instead of latest npm version.
+  - Added tests for lib/couchbase/client.js
+
+- 0.0.1 2013-08-17
   Initial release.